Create Student's_Grades.java
This is a Java program that prompts a user to input student's marks in 7 different units (Math, English, Chemistry, Biology, Computer, Physics and Geography), and the program stores the marks entered in an array.
The program Computes the sum, average, minimum and maximum grades for the particular student and determines whether the student has qualified to move on to the next class. Note, they are eligible if they have a minimum of B.
Average Mark	Grade:
70 and above : A
60-69 : B
50-59 : C
<50	Fail

Finally, the program computes the range of the numbers in the array.